+++ spec/lib/kb/cve_2014_3483_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 describe "The CVE-2014-3483 vulnerability" do
 	before(:all) do
 		@check = Dawn::Kb::CVE_2014_3483.new
-		@check.debug = true
+		# @check.debug = true
 	end
   it "is reported when a rails gem version 4.0.6 is detected" do
     @check.dependencies = [{:name=>"rails", :version=>"4.0.6"}]