+++ Changelog.md
@@ -5,7 +5,7 @@ It supports [Sinatra](http://www.sinatrarb.com),
 [Padrino](http://www.padrinorb.com) and [Ruby on Rails](http://rubyonrails.org)
 frameworks.
 
-_latest update: Tue Nov  1 22:47:56 CET 2016_
+_latest update: Fri Nov 11 16:20:33 CET 2016_
 
 * Removed CVSS as external dependency. Now it is a Dawn::Cvss module
 * Dropped sys-uname dependency. Using shell 'uname -r' instead
@@ -14,6 +14,7 @@ _latest update: Tue Nov  1 22:47:56 CET 2016_
 * Applied pull request #216 - Add TraviCI setup to README - by @vpolimenov
 * Applied pull request #215 - Fixes issue where debug mode was always enabled by @bartj3
 * Applied pull request #214 - Fixes copy in OWAP RoR mitigation message by bartj3
+* Adding a check for CVE-2016-3693: Information disclosure for safemode gem
 
 
 ## Version 1.6.6 - codename: Tow Mater (2016-11-01)